Move all project code to flanders package
