+++ flanders/__init__.py
@@ -1,0 +1,1 @@
+"""Modules defining FlandersBOT and its launch behaviour."""
