+++ flanders/cogs/__init__.py
@@ -1,0 +1,1 @@
+"""All cogs/extensions used by Flanders."""
