+++ flanders/cogs/futurama.py
@@ -2,7 +2,7 @@ import discord
 from compuglobal.aio import Morbotron
 from discord import app_commands
 
-from cogs._tvshows import TVShowCog
+from flanders.cogs._tvshows import TVShowCog
 
 
 class Futurama(TVShowCog):
